+++ packages/zod-schemas/src/index.ts
@@ -1,2 +1,4 @@
+export * from "./arrays";
 export * from "./principal";
+export * from "./schema-id";
 export * from "./url";