+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 | `@icp-sdk/canisters`   | v3.5.1  | Maintained ⚙ |
 | `@dfinity/zod-schemas` | v3.1.0  | Enhanced 🔧️  |
 
+### Features
+
+- Add `PrincipalSchema` and `Uint8ArraySchema` to Zod custom schemas.
+- Identify schemas with a metadata `id` corresponding to the enum `ZodSchemaId`.
+
 ## 93
 
 ### Overview